Added functionality to check for water within range of Flora Added functionality to reproduce flora when waterLevel is sufficient Added WaterSource objects Added Flora prefab

+++ Ecosystem Simulator/Assets/Scripts/Predator.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class Predator : MonoBehaviour
 {
     int nutrientLevel;
-    int waterLevel;
+    double waterLevel;
 
     int maxHealth;
     int currentHealth;
